+++ BE-dev/Orchid-BE/src/main/java/org/jio/orchidbe/models/auctions/Auction.java
@@ -82,7 +82,7 @@ public class Auction extends BaseEntity {
    private LocalDateTime remindAt;
 
    @Column(name = "bidding_price")
-   private Float biddingPrice;
+   private Double biddingPrice;
 
     @Column(name = "description")
     private String description;
