+++ BE-dev/Orchid-BE/src/main/java/org/jio/orchidbe/requests/bids/UpdateBiddingRequest.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @Builder
 public class UpdateBiddingRequest {
     private Long userId;
-    private Float biddingPrice;
+    private Double biddingPrice;
 }
